refactor: remove interactive mascot and global keyboard monitoring to implement a lightweight, centered TouchBar quota view.
